+++ public/administrate/js/components/searchbar.js
