+++ public/administrate/js/ptUserManagement.js
@@ -1,6 +1,6 @@
-import { loadSearchBar } from '/administrate/js/searchbar.js';
-import { loadFilter } from '/administrate/js/filter.js';
-import { loadSearchDropdown } from '/administrate/js/searchDropdown.js';
+import { loadSearchBar } from '/administrate/js/components/searchbar.js';
+import { loadFilter } from '/administrate/js/components/filter.js';
+import { loadSearchDropdown } from '/administrate/js/components/searchDropdown.js';
 
 $(document).ready(function () {
     loadContent();
@@ -11,15 +11,17 @@ function loadContent() {
 
     let ptUserManagementHTML = `
         <div class="title">PT 회원 관리</div>
-            <div id="searchTool">
-                <div id="searchDropdown"></div>
-                <div class="searchbar" data-placeholder="회원명 검색"></div>
-            </div>
-            
-            <div class="filter-group">
-                <div id="filter" data-type="1"></div>
-                <div id="filter" data-type="3"></div>       
+        <div id="searchTool">
+            <div id="searchDropdown"></div>
+            <div class="searchbar" data-placeholder="회원명 검색"></div>
+        </div>
+        
+        <div class="filter-group">
+            <div id="filter" data-type="1"></div>
+            <div id="filter" data-type="3"></div>       
         </div>      
+
+        <div id="ptUserTable"></div>
     `;
 
     container.html(ptUserManagementHTML);
